+++ CADCOP/src/Messages/MsgAMDLSColor.java
@@ -4,7 +4,7 @@ import AgentsAbstract.NodeId;
 
 public class MsgAMDLSColor extends MsgAMDLS {
 	private Integer color;
-	public MsgAMDLSColor(NodeId sender, NodeId reciever, Object context, int timeStamp, int time, int counter, Integer color) {
+	public MsgAMDLSColor(NodeId sender, NodeId reciever, Object context, int timeStamp, long time, int counter, Integer color) {
 		super(sender, reciever, context, timeStamp, time, counter);
 		this.color = color;
 	}
