+++ CADCOP/src/Messages/MsgAlgorithmFactor.java
@@ -4,7 +4,7 @@ import AgentsAbstract.NodeId;
 
 public class MsgAlgorithmFactor extends MsgAlgorithm{
 	
-    public MsgAlgorithmFactor(NodeId senderNode, NodeId recieverNode, Object context, int timeStamp, int time) {
+    public MsgAlgorithmFactor(NodeId senderNode, NodeId recieverNode, Object context, int timeStamp, long time) {
 		super(senderNode, recieverNode, context, timeStamp, time);
 		
 	}
