+++ CADCOP/src/Messages/MsgMgm2Phase3LR.java
@@ -4,7 +4,7 @@ import AgentsAbstract.NodeId;
 
 public class MsgMgm2Phase3LR extends MsgAlgorithm{
 
-	public MsgMgm2Phase3LR(NodeId sender, NodeId reciever, Object context, int timeStamp, int time) {
+	public MsgMgm2Phase3LR(NodeId sender, NodeId reciever, Object context, int timeStamp, long time) {
 		super(sender, reciever, context, timeStamp, time);
 	}
