+++ CADCOP/src/Messages/MsgMgm2Phase5IsBestLR.java
@@ -4,7 +4,7 @@ import AgentsAbstract.NodeId;
 
 public class MsgMgm2Phase5IsBestLR extends MsgAlgorithm{
 
-	public MsgMgm2Phase5IsBestLR(NodeId sender, NodeId reciever, Object context, int timeStamp, int time) {
+	public MsgMgm2Phase5IsBestLR(NodeId sender, NodeId reciever, Object context, int timeStamp, long time) {
 		super(sender, reciever, context, timeStamp, time);
 	}
