+++ CADCOP/src/Messages/MsgValueAssignmnet.java
@@ -4,7 +4,7 @@ import AgentsAbstract.NodeId;
 
 public class MsgValueAssignmnet extends MsgAlgorithm {
 
-	public MsgValueAssignmnet(NodeId sender, NodeId reciever, Object context, int timeStamp,int time) {
+	public MsgValueAssignmnet(NodeId sender, NodeId reciever, Object context, int timeStamp,long time) {
 		super(sender, reciever, context, timeStamp,time);
 	}
